Pin js-interpreter to version 2.2.0
There's an incompatible version 2.3.0 of js-interpreter currently,
which generates the following error at runtime:
`t.nativeGlobal.String is not a function`
which comes from the Interpreter.prototype.initString

For now, we're going to sidestep this by using the version we last
released with. In the long run, we should address the actual issue.