There are 3 changes here:
blockly was not working on node v12.
We have forked it into binary-com/blockly and cherry-pick the upstream fix.
In the long run, we should upgrade to the latest npm version for blockly.

circleci config has been updated to use the same node version (v12.13.0).

js-interperter package has breaking changes in 2.3.0,
therfore it's now fixed to "2.2.0" instead of "^2.2.0".