Fix blockly 1 building with gulp 4
We had been pinning our version of blockly to a version sha.
However now gulp 3 wouldn't build with node 12.

This commit pins it to our own forked version of blockly,
until we get to upgrade to the latest.

This does a fresh build of npm install and yarn import on
`node:12.18.0-buster` but this should then build okay on the circleci
version