Fix building in circleci and pin js-interpreter to 2.2.0 (#3287)
* Pin js-interpreter to version 2.2.0

There's an incompatible version 2.3.0 of js-interpreter currently,
which generates the following error at runtime:
`t.nativeGlobal.String is not a function`
which comes from the Interpreter.prototype.initString

For now, we're going to sidestep this by using the version we last
released with. In the long run, we should address the actual issue.

* Fix inconsistent circleci images and failing build

We were building with different versions of node for actual releases
and dev / staging. We should be consistent always so that the artifcacts
we deploy are the same ones we actually tested.

We're also moving from circleci: to cimg: because they deprecated
the old images and asked us to move.

Finally, we're moving to 12.18 because we've seen this image trying to
build on 12.13 without a cache.
We think that builds have only been working because of the circleci
cache of node_modules.
It looks from the builds, like [this
one](https://app.circleci.com/pipelines/github/binary-com/binary-bot/405/workflows/cfe00f01-7ee4-45d5-a4ed-a618e657789f/jobs/426)
that it was getting blockly and gulp from the existing cache

See [this github issue in
blockly](https://github.com/google/blockly/issues/2582).

This is a stepping stone. We need to fix the dependencies, upgrade
blockly, consistently use yarn OR npm, and commit package-lock or
yarn.lock as we work.

This is just an incremental fix.

See also [exiting task](https://redmine.deriv.cloud/issues/8729).

* Generate a new package-lock

Generated with node:12.18.0-buster since that's the only one that
doesn't error

We're okay to replace the existing package-lock because it hasn't been
used to generate the production releases.

* Added yarn.lock file generated from package-lock

The build process in circleci currently uses yarn,
but without a lock file.

This commits a lock file from today's package-lock, giving us some
stability

* Fix blockly 1 building with gulp 4

We had been pinning our version of blockly to a version sha.
However now gulp 3 wouldn't build with node 12.

This commit pins it to our own forked version of blockly,
until we get to upgrade to the latest.

This does a fresh build of npm install and yarn import on
`node:12.18.0-buster` but this should then build okay on the circleci
version